Major refactor
Core of the code has not been changed but significant refactor around it.
One aim is to separate the Viewer from the Legend so they can be positioned and styled separately and provide a utility hook to join them together. This means I've introduces AnnotationView for the main panel and AnnotationViewer for the two together and useAnnotations for the hook.

In addition:

- upgrade storybook setup
- improve darkmode handling
- add option to use contrast text
- adds layout options to Legend and AnnotationViewer

Remain to
- Fix the example
- update the jsdoc and readme
- maybe, add some tests